requirements.txt
Flask==3.1.3
Jinja2==3.1.6
MarkupSafe==3.0.3
Werkzeug==3.1.6
blinker==1.9.0
certifi==2026.2.25
charset-normalizer==3.4.4
click==8.3.1
google-api-core==2.30.0
google-api-python-client==2.192.0
google-auth==2.49.0
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.3.0
httplib2==0.31.2
idna==3.11
itsdangerous==2.2.0
protobuf==6.33.5
pyasn1==0.6.2
pyasn1-modules==0.4.2
pyparsing==3.3.2
python-dotenv==1.1.0
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
urllib3==2.6.3
